Simplify Computation/GenericComputation/CompileInstance
These were all similiar classes;

One difference is the presence of 'compilation devices' and 'devices',
but it's not clear how these are used (not needed for TS at least),
so I'd rather start simple.

I tried to set the pattern that Lower(), Compile(), and Execute() all
use a 'Computation' object. The backend could choose to make different
specializations of Computation object (e.g. LoweredComputation,
CompiledComputation), or not.